source/requirements.txt
django-rest-framework
psycopg2
django-filter
yfinance
gunicorn